moves videoViewer
